Add very basic file upload functionality
We need to upload some job descriptions that were previously hosted by the old service manual repository.

There is currently only support for uploading images when editing guides, so we need a way to allow any file type to be uploaded.

This is intentionally a very bare bones implementation designed to solve a last minute problem – it’s not discoverable within the application other than going directly to the URL. In an ideal world, this would be a rake task or similar – but then we’d need a way to get the files onto the server where that task is which defeats the point somewhat.