Add change note migrator
This adds a `ChangeNoteMigrator` class, which is designed to enable us
to update the change notes on existing editions where en masse where
they've been populated with the wrong data erroneously.

In short, it updates change notes for an edition, or makes an edition
minor or major, then republishes via the `GuideRepublisher` class,
providing some pretty console output along the way.

Also included is an example rake task which can be used for running the
thing.