Only validate dirty attributes in ChangeNoteMigrator
This adds a (rather unpleasant) mechanism to only validate attributes of
an edition that we've actually changed.

There's a bunch of records in the production database that is technically
invalid – it doesn't contain a `created_by` user, for example. These
records are generally quite old, and weren't created against the current
database schema.

We don't want to disable validation altogether – there's some important
stuff that we need to check around change notes and update types.

Subsequently, this adds some custom validation logic that:
- Assigns the attributes to the model
- Validates, then iterates through the errors found to see if any are on
  keys that have been changed
- If there are errors, it re-runs validation and triggers an exception
- If there aren't, it saves without validation