Re-order fields on topic form
This re-orders the topic form to be consistent with the guide form – with title, path and then description. At present we still need to differentiate between slug (used on the guide form) and path (used on the topic form) because the Guide form is only asking for a slug for the title (my-first-guide) whereas the Topic form expects an entire path (/service-manual/my-first-topic).