Fix history of ‘Decide how to host your service’
This guide has a troubled history, from a bug in the publisher that allowed users to save changes even if the underlying data had changed – e.g. if they had a guide open in a window, and another user had made changes or published a guide, they could save a diverged version on top of it.

Both the first and second ‘versions’ have been published twice. The first version is also major, which means that ‘Guidance first published’ shows up twice in the front-end. To fix this, we rewrite its history and ‘create’ two new versions from the second publication in each case, bumping the other version numbers appropriately.