Publish homepage using exact route, not prefix
This fixes a bug introduced by a recent change [1] which means the content store redirects requests for routes within a prefix to their base path. In practise, this means that requests for non existent routes (e.g. /service-manual/nonexistent-page) are 303 redirected to the content store item for the homepage (/service-manual). The Service Manual frontend follows these redirects and then renders the homepage rather than returning a 404 as you’d expect.

We originally chose to use a prefix route [2] because it meant ‘that requests for anything starting with /service-manual will be routed to our application, so we’ll be able to see 404s etc in our logs.’ This might have made some sense previously, but this behaviour now exhibited outweighs any such benefit.

[1]: https://github.com/alphagov/content-store/pull/269
[2]: https://github.com/alphagov/service-manual-publisher/commit/0f1f3a4a051c0446103b53972e125b6d9631d8ca