Don’t lint in default task; update to match CI
govuk.buildProject runs linting as part of a separate step by calling `bundle exec govuk-lint-ruby` directly, so we shouldn’t also run it as part of the default task.

This also updates the lint task to bring it in line with the arguments buildProject uses, to provide consistent results between local and CI linting.