Update govuk_app_config
[govuk_app_config changelog](https://github.com/alphagov/govuk_app_config/blob/master/CHANGELOG.md)
which contains details of what to update.

By updating this gem and the associated config, we'll get the standard GOV.UK
settings for logstasher, statsd, unicorn and sentry.

In particular, this update sets default logstasher config, so that the
[deployment dashboard](https://grafana.publishing.service.gov.uk/dashboard/file/deployment_service-manual-publisher.json?refresh=5s&orgId=1)
for this app will now work.  This is a first port of call for support engineers
troubleshooting problems.

It also helps ensure that there's a consistent feel across applications when we
look at [Kibana](https://reliability-engineering.cloudapps.digital/manuals/logit-io-joiners-and-leavers.html#content)
for application logs.