Remove multiple templates as mail-notify supports only single body parts
Previously we sent both a text and an html part; however, mail-notify
will send only the body part, which is blank if there are multiple
alternative parts.

I've settled on sending the HTML part, as Notify (or perhaps
mail-notify?) will pass this through Markdown and send HTML anyway;
however, if sticking with the text part is preferable we can do that
instead.