DRY-up republishing the homepage
Previously republishing the homepage would result in a "minor" update,
which is inaccurate in a "republish" task. The task also did not update
the links for homepage, nor does any other code. This rewrites the task
to use the new Republisher, and changes the presenter to explicitly say
the homepage has no links (or at least, none that we are controlling).