Support other kinds of homepage publishing
Previously we DRYed-up the "republish:homepage" task to use the new
Republisher, which sets "republish" as the update type. This adds
support for an optional "UPDATE_TYPE" environment variable when running
the task, so that we can still use it to publish minor (or even major)
content changes on the service manual homepage.