Temporary hack to delete email-signup links
Previously we stopped sending "email_alert_signup" links to the
Publishing API [1]. However, this does not cause the existing links
to be deleted. Instead, we need to explicitly send an empty array
in the links payload, republish, and then remove it altogether. This
sets an empty array temporarily, to be removed after republishing.

[1]: 11a8a67