Switch to db/schema.rb
Previously running a new migration with db/structure.sql seemed to
regenerate most of the file, which isn't great for the reviewer of
the changes. Rather than try to understand equivalence between the
old and new versions, this (mostly) replaces it with db/schema.rb

It turns out that the new schema format also works with RuboCop, but
since this is meant to be a refactor, we should avoid retrospectively
changing the DB to satisfy linting issues (hence disabling the Cop).