Remove Jasmine Helper
This was added 4 years ago:
https://github.com/alphagov/service-manual-publisher/commit/7728c88cda58c2cebdacf63cd05ebf0cbd3b4ec2#diff-54dc43b93174691dc226df624651d227R1-R4

We explicitly stopped PhantomJS from being downloaded, to use the
PhantomJS that is already on the PATH. But it's not clear why we
did this (nor is it clear from the PR description in #129).
I assume it was just copied from signon (or another app), which
has had the helper for 6 years:
https://github.com/alphagov/signon/commit/04e34a8eef138e99d92344178d3ebc60c344a005#diff-54dc43b93174691dc226df624651d227R1-R4

This helper has been working fine, but recently the CI build has
broken, whereas the same code works fine in Docker. So the less
environmental manipulation we brittly rely on, the better.

Using the PhantomJS included with Jasmine seems to fix the build,
so we no longer need the helper file.

This is the same approach we have taken as in Signon, which was
plagued with the same issue:
https://github.com/alphagov/signon/pull/1472/commits/4a0410890e2342b94283fc6627d23943e920b38d