Use == for checking XMLHttpRequest status
This === was causing a problem here as the test mocks this as a string,
whereas I'm relatively confident this is an integer in the real world.
However I'm a bit confused about where currentTarget comes into play so
I found myself a bit confused.

The simplest way to resolve this for now is to disable this line.