Specify version of ActiveRecord::Migration to inherit from
Running `make service-manual-publisher` in govuk-docker is failing
with this error:

> Directly inheriting from ActiveRecord::Migration is not supported.
> Please specify the Rails release the migration was written for

In June, we made a change to govuk-docker which prevents errors
like this from being swallowed up:
https://github.com/alphagov/govuk-docker/commit/b6eb8eb517fd18e9d97a0b7e6d4e0f2eef3f47e1#diff-2e9ef33560da433746fdb218886c1924da9cb7c85d2e7e3f5f667d21a5126c3a

Therefore we now need to fix the errors in order to build the
Docker container.

I've chosen the version 5.2 as this is the one specified in the
most recent migration in 264d853acd89b4eb43160cbb84d1476369baa953.