Move Diffy::CSS's dynamic styling into edition change show view
Dart Sass doesn't compile scss.erb files. "Diffy::CSS" returns styling.
There's a couple of ways to resolve this.

1. we could move that styling form a dynamic call into the diffy stylesheet

This would mean that were Diffy:CSS to update it's styling we would lose
any changes. This feels fairly unlikely.

2. we move the dynamic call to a style tag inside the body.

This isn't optimum for speed as it doesn't create a new request for the
additional styling. However, this is only used on this 1 page and the
page load time difference is minimal.

This uses the second option.