Set JSON encoding so timezone outputs to json will not have trailing milliseconds
By default, timezone to json string outputs an standard ISO 8601 formatted datetime along with '.ooo' milliseconds.
While using this as_json method in the tests, it is failing because the application is outputting datetime format without the milliseconds. To make it consistent, we should also just set this globally so we are outputting the same thing, but it might be worth reconsidering given the .000 format is standard.