Temporarily skip test
Due to an issue with, most likely, the webdriver, the tests fail in the CI
with the error: "Unable to find modal dialog". This is specific to chrome
version 127.0.6533.72.
Currently all tests are passing locally, as the docker image is
pulling an older version of `120.0.6099.224' for the browser and driver.

Temporarily skipping the tests to unblock merging in pending PRs.
The application and the pop-ups have been manually tested and work as expected.