Updating tests to look at email body string instead of individual parts.
Looks like previously there were no parts to the email and hence the assertions were not being run. With the update, first part had content, but the subsequent parts had no body, which led to the failures. We can avoid this by testing directly on the email body rather than looking at the parts.